+++ packages/convert/app/components/use-probe.ts
@@ -16,71 +16,6 @@ import {Source} from '~/lib/convert-state';
 
 export type ProbeResult = ReturnType<typeof useProbe>;
 
-export const useThumbnail = ({
-	src,
-	logLevel,
-	onVideoThumbnail,
-}: {
-	src: Source;
-	logLevel: LogLevel;
-	onVideoThumbnail: (videoFrame: VideoFrame) => void;
-}) => {
-	const abortController = new AbortController();
-	parseMedia({
-		signal: abortController.signal,
-		reader: src.type === 'file' ? webFileReader : fetchReader,
-		src: src.type === 'file' ? src.file : src.url,
-		logLevel,
-		onVideoTrack: async (track) => {
-			if (typeof VideoDecoder === 'undefined') {
-				return null;
-			}
-
-			let frames = 0;
-
-			const decoder = new VideoDecoder({
-				error: (error) => {
-					// eslint-disable-next-line no-console
-					console.log(error);
-				},
-				output(frame) {
-					onVideoThumbnail(frame);
-					frame.close();
-				},
-			});
-
-			if (!(await VideoDecoder.isConfigSupported(track)).supported) {
-				return null;
-			}
-
-			// TODO: See if possible
-			decoder.configure(track);
-			return (sample) => {
-				frames++;
-				if (frames >= 30) {
-					abortController.abort();
-				}
-
-				decoder.decode(new EncodedVideoChunk(sample));
-			};
-		},
-	}).catch((err) => {
-		if ((err as Error).stack?.includes('Cancelled')) {
-			return;
-		}
-		if ((err as Error).stack?.toLowerCase()?.includes('aborted')) {
-			return;
-		}
-		// firefox
-		if ((err as Error).message?.toLowerCase()?.includes('aborted')) {
-			return;
-		}
-
-		// eslint-disable-next-line no-console
-		console.log(err);
-	});
-};
-
 export const useProbe = ({
 	src,
 	logLevel,
